+++ block-buffer/src/sealed.rs
@@ -1,8 +1,15 @@
 use super::{ArrayLength, Block};
-use core::slice;
+use core::{ops::Sub, slice};
+use generic_array::typenum::{Gr, IsGreater, IsLess, Le, NonZero, Unsigned, B1, U1, U256};
 
 /// Sealed trait for buffer kinds.
 pub trait Sealed {
+    type Pos: Default + Clone;
+
+    fn get_pos(buf: &[u8], pos: &Self::Pos) -> usize;
+
+    fn set_pos(buf_val: &mut [u8], pos: &mut Self::Pos, val: usize);
+
     /// Invariant guaranteed by a buffer kind, i.e. with correct
     /// buffer code this function always returns true.
     fn invariant(pos: usize, block_size: usize) -> bool;
@@ -12,6 +19,17 @@ pub trait Sealed {
 }
 
 impl Sealed for super::Eager {
+    type Pos = ();
+
+    fn get_pos(buf: &[u8], _pos: &Self::Pos) -> usize {
+        buf[buf.len() - 1] as usize
+    }
+
+    fn set_pos(buf: &mut [u8], _pos: &mut Self::Pos, val: usize) {
+        debug_assert!(val <= u8::MAX as usize);
+        buf[buf.len() - 1] = val as u8;
+    }
+
     #[inline(always)]
     fn invariant(pos: usize, block_size: usize) -> bool {
         pos < block_size
@@ -36,6 +54,17 @@ impl Sealed for super::Eager {
 }
 
 impl Sealed for super::Lazy {
+    type Pos = u8;
+
+    fn get_pos(_buf_val: &[u8], pos: &Self::Pos) -> usize {
+        *pos as usize
+    }
+
+    fn set_pos(_buf_val: &mut [u8], pos: &mut Self::Pos, val: usize) {
+        debug_assert!(val <= u8::MAX as usize);
+        *pos = val as u8;
+    }
+
     #[inline(always)]
     fn invariant(pos: usize, block_size: usize) -> bool {
         pos <= block_size
@@ -64,4 +93,14 @@ impl Sealed for super::Lazy {
             )
         }
     }
+}
+
+pub trait BlockSizes {}
+
+impl<T: Unsigned> BlockSizes for T
+where
+    Self: IsLess<U256> + IsGreater<U1> + Sub<B1>,
+    Le<Self, U256>: NonZero,
+    Gr<Self, U1>: NonZero,
+{
 }